client.rt_info() fails to run HTTP GET request
In the current state, the rt_info() method errors out with a stack trace
on the call to self.sess.get().

This function call is trying to reach "rt" without the rest of the base
URI.

This means that other methods that are calling rt_info() to get their
information, namely rt_version() and rt_plugins(), are broken too.

By reusing the _get() method we can fix and simplify rt_info().